refactor: separate validation code
